[atlas-gradle-plugin] 可能是一个没有资源的awb

+++ atlas-gradle-plugin/atlas-plugin/src/main/java/com/android/builder/core/AtlasBuilder.java
@@ -364,7 +364,8 @@ public class AtlasBuilder extends AndroidBuilder {
 
                 getTargetInfo();
 
-                AaptOptions aaptOptions = (AaptOptions)FieldUtils.readField(aaptConfigBuilder.build(), "mAaptOptions",
+                AaptOptions aaptOptions = (AaptOptions)FieldUtils.readField(aaptConfigBuilder.build(),
+                                                                            "mAaptOptions",
                                                                             true);
 
                 ProcessExecutor processExecutor = (ProcessExecutor)FieldUtils.readField(aapt, "mProcessExecutor", true);
@@ -375,7 +376,11 @@ public class AtlasBuilder extends AndroidBuilder {
                 PngProcessMode processMode = (PngProcessMode)FieldUtils.readField(aapt, "mProcessMode", true);
                 int cruncherProcesses = aaptOptions.getCruncherProcesses();
 
-                aapt = new AtlasAapt(processExecutor, processOutputHandler, buildToolInfo, getLogger(), processMode,
+                aapt = new AtlasAapt(processExecutor,
+                                     processOutputHandler,
+                                     buildToolInfo,
+                                     getLogger(),
+                                     processMode,
                                      cruncherProcesses);
             } catch (Throwable e) {
                 throw new GradleException("aapt exception", e);
@@ -449,9 +454,13 @@ public class AtlasBuilder extends AndroidBuilder {
 
             if (hasResources) {
                 // Load the main symbol file.
-                File explodedDir = atlasExtension.getProject().file(
-                    atlasExtension.getProject().getBuildDir().getAbsolutePath() + "/" + FD_INTERMEDIATES
-                        + "/exploded-ap" + "/");
+                File explodedDir = atlasExtension.getProject().file(atlasExtension.getProject()
+                                                                        .getBuildDir()
+                                                                        .getAbsolutePath()
+                                                                        + "/"
+                                                                        + FD_INTERMEDIATES
+                                                                        + "/exploded-ap"
+                                                                        + "/");
                 File mainRTxt2 = new File(explodedDir, "R.txt");
                 SymbolTable mainSymbols2 = mainRTxt2.isFile() ? SymbolIo.read(mainRTxt2)
                     : SymbolTable.builder().build();
@@ -471,8 +480,9 @@ public class AtlasBuilder extends AndroidBuilder {
                     throw new IOException("Failed to read manifest " + depMan.getAbsolutePath(), e);
                 }
                 if (mainPackageName.equals(depPackageName) && enforceUniquePackageName) {
-                    throw new RuntimeException(
-                        String.format("Error: A library uses the same package as this project: %s", depPackageName));
+                    throw new RuntimeException(String.format(
+                        "Error: A library uses the same package as this project: %s",
+                        depPackageName));
                 }
                 File rFile = dependency.getSymbolFile();
                 SymbolTable depSymbols = rFile.exists() ? SymbolIo.read(rFile) : SymbolTable.builder().build();
@@ -618,7 +628,8 @@ public class AtlasBuilder extends AndroidBuilder {
         checkState(getTargetInfo() != null, "Cannot call processResources() before setTargetInfo() is called.");
 
         // launch aapt: create the command line
-        ProcessInfo processInfo = aaptCommand.build(getTargetInfo().getBuildTools(), getTargetInfo().getTarget(),
+        ProcessInfo processInfo = aaptCommand.build(getTargetInfo().getBuildTools(),
+                                                    getTargetInfo().getTarget(),
                                                     getLogger());
 
         processInfo = new TProcessInfo(processInfo, aaptCommand.getSymbolOutputDir());
@@ -655,17 +666,21 @@ public class AtlasBuilder extends AndroidBuilder {
                 FileUtils.copyFile(mainSymbolFile, mergedSymbolFile);
             }
 
-            FileUtils.writeLines(mergedSymbolFile, FileUtils.readLines(awbSymbolFile), true);
+            if (awbSymbolFile.exists()) {
+                FileUtils.writeLines(mergedSymbolFile, FileUtils.readLines(awbSymbolFile), true);
+            }
             if (incrementalBuild) {
                 if (baseSymbolFile != null) {
                     FileUtils.writeLines(mergedSymbolFile, FileUtils.readLines(baseSymbolFile), true);
                 } else {
-                    FileUtils.writeLines(mergedSymbolFile, Collections2
-                        .filter(FileUtils.readLines(mergedSymbolFile), input -> !input.contains("styleable")));
+                    FileUtils.writeLines(mergedSymbolFile,
+                                         Collections2.filter(FileUtils.readLines(mergedSymbolFile),
+                                                             input -> !input.contains("styleable")));
 
                     sLogger.warn(String.format(
                         "Please check and update your baseline project atlasplugin : processAwbResources expected a base symbol file"
-                            + " at:\n %s \nbut the file does not exist or is a directory", baseSymbolFile));
+                            + " at:\n %s \nbut the file does not exist or is a directory",
+                        baseSymbolFile));
                 }
             }
         } catch (IOException e) {
@@ -955,7 +970,8 @@ public class AtlasBuilder extends AndroidBuilder {
 
                     multiDexer.dexMerge(dexs, outDexFolder);
                 } else {
-                    DexMerger dexMerger = new DexMerger(dexs.toArray(new Dex[0]), CollisionPolicy.KEEP_FIRST,
+                    DexMerger dexMerger = new DexMerger(dexs.toArray(new Dex[0]),
+                                                        CollisionPolicy.KEEP_FIRST,
                                                         new DxContext());
                     Dex dex = dexMerger.merge();
 
@@ -972,7 +988,8 @@ public class AtlasBuilder extends AndroidBuilder {
 
             long finishTime = System.currentTimeMillis();
 
-            sLogger.info("[mtldex] dex consume {} , dexmerge consume {}", endDexTime - startTime,
+            sLogger.info("[mtldex] dex consume {} , dexmerge consume {}",
+                         endDexTime - startTime,
                          finishTime - endDexTime);
         } else {
             super.convertByteCode(inputs, outDexFolder, multidex, mainDexList, dexOptions, processOutputHandler);
@@ -1015,8 +1032,8 @@ public class AtlasBuilder extends AndroidBuilder {
         String md5 = "";
         File dexFile = new File(outFile, "classes.dex");
 
-        if (!inputFile.getName().startsWith("combined") && !(inputFile.getName().startsWith("main") && inputFile
-            .getName().endsWith("jar"))) {
+        if (!inputFile.getName().startsWith("combined") && !(inputFile.getName().startsWith("main")
+            && inputFile.getName().endsWith("jar"))) {
 
             if (inputFile.isFile()) {
                 md5 = MD5Util.getFileMD5(inputFile);
@@ -1027,7 +1044,8 @@ public class AtlasBuilder extends AndroidBuilder {
             if (StringUtils.isNotEmpty(md5)) {
                 AtlasBuildContext.sBuilderAdapter.fileCache.fetchFile(md5, dexFile, "pre-dex");
                 if (dexFile.exists() && dexFile.length() > 0) {
-                    sLogger.info("[mtldex] hit cache dex for {} , {}", inputFile.getAbsolutePath(),
+                    sLogger.info("[mtldex] hit cache dex for {} , {}",
+                                 inputFile.getAbsolutePath(),
                                  dexFile.getAbsolutePath());
                     return;
                 } else {
@@ -1107,14 +1125,16 @@ public class AtlasBuilder extends AndroidBuilder {
             if (jniLoc.isFile()) {
                 jniArchiveMods.put(jniLoc, FileStatus.NEW);
             } else {
-                Set<RelativeFile> files = RelativeFiles.fromDirectory(jniLoc, RelativeFiles
-                    .fromPathPredicate(nativeLibraryPredicate));
+                Set<RelativeFile> files = RelativeFiles.fromDirectory(jniLoc,
+                                                                      RelativeFiles.fromPathPredicate(
+                                                                          nativeLibraryPredicate));
                 jniMods.putAll(Maps.asMap(files, Functions.constant(FileStatus.NEW)));
             }
         }
 
         Set<RelativeFile> assets = assetsFolder == null ? Collections.emptySet() : RelativeFiles.fromDirectory(
-            assetsFolder, new java.util.function.Predicate<RelativeFile>() {
+            assetsFolder,
+            new java.util.function.Predicate<RelativeFile>() {
                 @Override
                 public boolean test(RelativeFile relativeFile) {
                     return relativeFile.getFile().isFile();
@@ -1126,14 +1146,19 @@ public class AtlasBuilder extends AndroidBuilder {
         boolean v1SigningEnabled = false;
         boolean v2SigningEnabled = false;
 
-        ApkCreatorFactory.CreationData creationData = new ApkCreatorFactory.CreationData(outApkLocation, key,
-                                                                                         certificate, v1SigningEnabled,
-                                                                                         v2SigningEnabled, null,
+        ApkCreatorFactory.CreationData creationData = new ApkCreatorFactory.CreationData(outApkLocation,
+                                                                                         key,
+                                                                                         certificate,
+                                                                                         v1SigningEnabled,
+                                                                                         v2SigningEnabled,
+                                                                                         null,
                                                                                          // BuiltBy
-                                                                                         "atlas", minSdkVersion,
+                                                                                         "atlas",
+                                                                                         minSdkVersion,
                                                                                          NativeLibrariesPackagingMode.COMPRESSED,
                                                                                          noCompressPredicate::apply);
-        try (OldPackager packager = new OldPackager(creationData, androidResPkgLocation,
+        try (OldPackager packager = new OldPackager(creationData,
+                                                    androidResPkgLocation,
                                                     LoggerWrapper.getLogger(AtlasBuilder.class))) {
             // add dex folder to the apk root.
             if (!dexFolders.isEmpty()) {
@@ -1146,7 +1171,8 @@ public class AtlasBuilder extends AndroidBuilder {
             }
 
             for (Map.Entry<File, FileStatus> resourceArchiveUpdate : javaResourceArchiveMods.entrySet()) {
-                packager.updateResourceArchive(resourceArchiveUpdate.getKey(), resourceArchiveUpdate.getValue(),
+                packager.updateResourceArchive(resourceArchiveUpdate.getKey(),
+                                               resourceArchiveUpdate.getValue(),
                                                new java.util.function.Predicate<String>() {
                                                    @Override
                                                    public boolean test(String s) {
@@ -1160,7 +1186,8 @@ public class AtlasBuilder extends AndroidBuilder {
             }
 
             for (Map.Entry<File, FileStatus> resourceArchiveUpdate : jniArchiveMods.entrySet()) {
-                packager.updateResourceArchive(resourceArchiveUpdate.getKey(), resourceArchiveUpdate.getValue(),
+                packager.updateResourceArchive(resourceArchiveUpdate.getKey(),
+                                               resourceArchiveUpdate.getValue(),
                                                new java.util.function.Predicate<String>() {
                                                    @Override
                                                    public boolean test(String s) {
@@ -1182,14 +1209,18 @@ public class AtlasBuilder extends AndroidBuilder {
 
     private static final String MANIFEST_TEMPLATE = "<?xml version=\"1.0\" encoding=\"utf-8\"?>\n"
         + "<manifest xmlns:android=\"http://schemas.android.com/apk/res/android\"\n"
-        + "          package=\"${packageName}\"\n" + "    xmlns:tools=\"http://schemas.android.com/tools\"\n"
-        + "          android:versionCode=\"${versionCode}\"\n" + "          android:versionName=\"${versionName}\">\n"
-        + "    <application></application>\n" + "</manifest>\n";
+        + "          package=\"${packageName}\"\n"
+        + "    xmlns:tools=\"http://schemas.android.com/tools\"\n"
+        + "          android:versionCode=\"${versionCode}\"\n"
+        + "          android:versionName=\"${versionName}\">\n"
+        + "    <application></application>\n"
+        + "</manifest>\n";
 
     public void mergeManifestsForBunlde(File outputFile, String packageName, String versionName, String versionCode)
         throws IOException {
 
-        String xml = MANIFEST_TEMPLATE.replace("${packageName}", packageName).replace("${versionCode}", versionCode)
+        String xml = MANIFEST_TEMPLATE.replace("${packageName}", packageName)
+            .replace("${versionCode}", versionCode)
             .replace("${versionName}", versionName);
 
         outputFile.getParentFile().mkdirs();
@@ -1204,8 +1235,10 @@ public class AtlasBuilder extends AndroidBuilder {
             return super.getDexByteCodeConverter();
         }
         if (dexByteCodeConverter == null) {
-            dexByteCodeConverter = new DexByteCodeConverterHook(getLogger(), defaultBuilder.getTargetInfo(),
-                                                                javaProcessExecutor, verboseExec);
+            dexByteCodeConverter = new DexByteCodeConverterHook(getLogger(),
+                                                                defaultBuilder.getTargetInfo(),
+                                                                javaProcessExecutor,
+                                                                verboseExec);
         }
         sLogger.debug("use DexByteCodeConverterHook......");
         return dexByteCodeConverter;
